+++ docs/src/lib/generate-content-hash.js
@@ -1,13 +1,17 @@
 const fs = require('fs');
 const glob = require('fast-glob');
-const md5 = require('md5');
+const crypto = require('node:crypto');
 
 function generateContentHash(dir) {
   const sourceFiles = glob.sync([`${dir}`]);
   const sourceContent = sourceFiles
     .map((sourceFile) => fs.readFileSync(sourceFile))
     .join('');
-  return md5(sourceContent).slice(0, 8);
+  return crypto
+    .createHash('md5')
+    .update(sourceContent)
+    .digest('hex')
+    .slice(0, 8);
 }
 
 module.exports = generateContentHash;